OCT-111: moved UpsertCatalogQuery
